# packages/utils/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quillhive API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    get:
      operationId: getGroups
      tags: [groups]
      summary: List groups/communities
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Groups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupsListResponse"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{id}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get a group by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{id}/join:
    post:
      operationId: joinGroup
      tags: [groups]
      summary: Join or leave a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Membership toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinGroupResponse"
  /groups/{id}/posts:
    get:
      operationId: getGroupPosts
      tags: [groups]
      summary: Get posts in a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Group posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsListResponse"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        isFollowing:
          type: boolean
      required:
        - id
        - username
        - email
        - displayName
        - createdAt
        - followersCount
        - followingCount
        - postsCount
        - isFollowing
...
    Post:
      type: object
      properties:
        id:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        title:
          type: ["string", "null"]
        content:
          type: string
        excerpt:
          type: ["string", "null"]
        type:
          type: string
          enum: [story, poem, novel, artwork, post, spark]
        visibility:
          type: string
          enum: [public, followers, private]
        imageUrl:
          type: ["string", "null"]
        attachments:
          type: array
          items:
            type: object
            additionalProperties: true
        tags:
          type: array
          items:
            type: string
        likesCount:
          type: integer
        commentsCount:
          type: integer
        isLiked:
          type: boolean
        isPublished:
          type: boolean
        visibility:
          type: string
          enum: [public, followers, private]
        groupId:
          type: ["integer", "null"]
        attachments:
          type: array
          items:
            type: object
            additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - authorId
        - author
        - content
        - type
        - tags
        - likesCount
        - commentsCount
        - isLiked
        - isPublished
        - createdAt
        - updatedAt
...
    PostsListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - posts
        - total
        - page
        - limit
...
    Group:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        category:
          type: string
        membersCount:
          type: integer
        postsCount:
          type: integer
        isMember:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - membersCount
        - postsCount
        - isMember
        - createdAt
    CreateGroupRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        privacy:
          type: string
          enum: [open, private]
        rules:
          type: ["string", "null"]
      required:
        - name
        - category
        - privacy
    GroupsListResponse:
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/Group"
        total:
          type: integer
        page:
          type: integer
      required:
        - groups
        - total
        - page
    JoinGroupResponse:
      type: object
      properties:
        isMember:
          type: boolean
        membersCount:
          type: integer
      required:
        - isMember
        - membersCount
...